requirements.txt
fastapi==0.97.0
uvicorn[standard]==0.22.0
pyjwt[crypto]==2.7.0
